Hacknet API update
Update to update to the new Hacknet API. Did not completely remove old code yet. Correction in daemon to have initial 100% values as well.

Not sure if it is all working properly with current version yet (v0.41.2)

Daemon also shows with errors in the in-game editor. Not sure if this is correct or not.